requirements.txt
python-dotenv
tweepy
nltk
selenium
konlpy
JPype1
matplotlib
pandas
numpy